Create Can chef
Chef owns a car that can run 
15
15 kilometers using 
1
1 liter of petrol.
He wants to attend a programming camp at DAIICT.

X
X liters of petrol is present in Chef's car. The distance between his house and DAIICT is 
Y
Y kilometers.
Determine whether Chef can attend the event and return to his home with the given amount of petrol.

Note: Chef has to return back to home, so the total distance to be covered would be 
2
⋅
Y
2⋅Y.

Input Format
The first line of input will contain a single integer 
T
T, denoting the number of test cases.
Each test case consists of two space-separated integers 
X
X and 
Y
Y — the amount of petrol in liter and the distance between Chef's house and DAIICT in kilometers, respectively.
Output Format
For each test case, print YES if it is possible for the chef to attend the event and return home, else, print NO.

You may print each character in uppercase or lowercase. For example, Yes, yes, YES, and yES are all considered identical.