Create Cyclic quadrilateral
You are given the sizes of angles of a simple quadrilateral (in degrees) 
A
A, 
B
B, 
C
C and 
D
D, in some order along its perimeter. Determine whether the quadrilateral is cyclic.

Note: A quadrilateral is cyclic if and only if the sum of opposite angles is 
18
0
∘
180 
∘
 .

Input
The first line of the input contains a single integer 
T
T denoting the number of test cases. The description of 
T
T test cases follows.
The first and only line of each test case contains four space-separated integers 
A
A, 
B
B, 
C
C and 
D
D.
Output
Print a single line containing the string "YES" if the given quadrilateral is cyclic or "NO" if it is not (without quotes).

You may print each character of the string in uppercase or lowercase (for example, the strings "yEs", "yes", "Yes" and "YES" will all be treated as identical).